+++ include/game/functions.h
@@ -1,6 +1,10 @@
 #define FUNCTION_H
 #ifdef FUNCTION_H
 
+#include <vector>
+#include "utils.h"
+
 float lerp(float min, float max, float blend);
+LevelRect getRect(LevelPos pos, LevelSize size);
 
 #endif