+++ src/map.cpp
@@ -3,7 +3,17 @@
 // INIT
 Map::Map(Game* g) : game(g) {
     loadData();
-    player = new Player(g, this);
+
+    BulletData bulletData = BulletData();
+    bulletData.damage = 1.0f;
+    bulletData.speed = 1.0f;
+
+    WeaponData weaponData = WeaponData();
+    weaponData.delay = 100;
+    weaponData.bullets = { bulletData };
+
+    Weapon* weapon = new Weapon(g, weaponData);
+    player = new Player(g, weapon);
 }
 
 void Map::loadData() {
@@ -33,17 +43,19 @@ void Map::render() {
     static H2DE_Engine* engine = game->getEngine();
     static Calculator* c = game->getCalculator();
     static GameData* gameData = game->getData();
-    static H2DE_Size itemSize = c->convertToPx({ 1.0f, 1.0f });
+
+    static LevelSize itemSize = gameData->sizes->item;
+    static H2DE_Size absItemSize = c->convertToPx(itemSize);
 
     for (Item* item : items) {
         H2DE_GraphicObject* i = H2DE_CreateGraphicObject();
         i->type = POLYGON;
-        i->pos = c->convertToPx(item->pos, { 1.0f, 1.0f });
+        i->pos = c->convertToPx(item->pos, itemSize);
         i->points = {
             { 0, 0 },
-            { itemSize.w, 0 },
-            { itemSize.w, itemSize.h },
-            { 0, itemSize.h },
+            { absItemSize.w, 0 },
+            { absItemSize.w, absItemSize.h },
+            { 0, absItemSize.h },
         };
         i->rgb = (item->type == WALL) ? H2DE_RGB{ 255, 255, 255, 255 } : (item->type == GROUND) ? H2DE_RGB{ 50, 50, 50, 255 } : H2DE_RGB{ 0, 0, 0, 255 };
         i->filled = true;