+++ CHANGES.md
@@ -1,3 +1,9 @@
+## HB-1.0.2
+#### ADDED
+- shoot delay
+- weapons
+- wall collisions
+
 ## HB-1.0.1
 #### ADDED
 - bullets shooting